refactor & fix bug for fixed price
